inventory/suppliers/urls.py
from django.urls import path

from . import views

app_name = "suppliers"

urlpatterns = [
    path("", views.SupplierListView.as_view(), name="list"),
    path("add/", views.SupplierCreateView.as_view(), name="add"),
]