pyproject.toml
[project]
name = "mbf-auto-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-browser-reload>=1.18.0",
    "django-cleanup>=9.0.0",
    "django-cotton>=1.5.3",
    "django-htmx>=1.22.0",
    "django-ninja>=1.3.0",
    "django-unicorn>=0.62.0",
    "djangosetupheist==0.6.2",
    "pillow>=11.1.0",
]
